renamed memberRepository to memberRepo for consistency
